# api/runtime.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClearSight Runtime API Contract",
    "version": "0.9.0",
    "description": "Executable production route, access-class and permission contract. Detailed request/response schemas remain in api/openapi.yaml until the typed-client tranche consolidates them."
  },
  "components": {
    "securitySchemes": {
      "signedIdentityEnvelope": { "type": "apiKey", "in": "header", "name": "X-ClearSight-Identity" },
      "signedIdentityTimestamp": { "type": "apiKey", "in": "header", "name": "X-ClearSight-Identity-Timestamp" },
      "signedIdentitySignature": { "type": "apiKey", "in": "header", "name": "X-ClearSight-Identity-Signature" },
      "captureSession": { "type": "http", "scheme": "bearer" }
    }
  },
  "security": [
    { "signedIdentityEnvelope": [], "signedIdentityTimestamp": [], "signedIdentitySignature": [] }
  ],
  "paths": {
...
    "/api/v1/evidence/session": { "get": { "operationId": "getEvidenceSession", "security": [{ "captureSession": [] }], "x-clearsight-route-class": "BOUNDED_CAPABILITY" } },
    "/api/v1/evidence/session/draft": {
      "get": { "operationId": "getEvidenceSessionDraft", "security": [{ "captureSession": [] }], "x-clearsight-route-class": "BOUNDED_CAPABILITY" },
      "put": { "operationId": "saveEvidenceSessionDraft", "security": [{ "captureSession": [] }], "x-clearsight-route-class": "BOUNDED_CAPABILITY" }
    },
    "/api/v1/evidence/session/submissions": { "post": { "operationId": "submitEvidenceSession", "security": [{ "captureSession": [] }], "x-clearsight-route-class": "BOUNDED_CAPABILITY" } },
    "/api/v1/evidence/artifacts": { "post": { "operationId": "uploadEvidenceArtifact", "security": [{ "signedIdentityEnvelope": [], "signedIdentityTimestamp": [], "signedIdentitySignature": [] }, { "captureSession": [] }], "x-clearsight-route-class": "AUTHENTICATED_OR_CAPABILITY" } },
...
  }
}